+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: pandalaf <pandalaf@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/25 19:26:47 by pandalaf          #+#    #+#             */
-/*   Updated: 2022/05/16 18:43:05 by pandalaf         ###   ########.fr       */
+/*   Updated: 2022/05/16 18:56:11 by pandalaf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int				ft_isascii(int ch);
 //Function determines whether input is a printing character.
 int				ft_isprint(int ch);
 //Function determines length of the string.
-unsigned int	ft_strlen(const char *str);
+size_t			ft_strlen(const char *str);
 //Function writes a value to a determined length of bytes in memory.
 void			ft_memset(void *byte, int ch, size_t len);
 //Function sets determined length of memory in a string to zero.
